Tidy up documentation remove unused crate references
