+++ Cargo.toml
@@ -1,20 +1,22 @@
 [package]
 name = "messaging_thread_pool"
-version = "0.1.0"
+version = "0.1.1"
 authors = ["cainem"]
 edition = '2021'
-description = "A library for a aiding the creation of a thread pool that communicates via channels"
+description = "A library for aiding the creation of a thread pool of objects that communicated with via channels"
 license = "MIT"
+homepage = "https://github.com/cainem/messaging-thread-pool"
+repository = "https://github.com/cainem/messaging-thread-pool"
+documentation = "https://docs.rs/messaging_thread_pool/latest/messaging_thread_pool/"
+readme = "README.md"
 
 [dependencies]
 tracing = { version = "0.1.36", features = [ "max_level_trace", "release_max_level_trace" ] }
 tracing-subscriber = { version = "0.3.15", features = [ "fmt", "std", "json"] }
 tracing-attributes = "0.1.22"
 tracing-appender = "0.2.2"
-backtrace = "0.3.66"
 crossbeam-channel = "0.5.6"
 rand = "0.8.5"
-arrayvec = "0.7.2"
 rand_xoshiro = "0.6.0"
 once_cell = "1.15.0"
