+++ src/id_provider/id_provider_mutex.rs
@@ -3,7 +3,8 @@ use std::sync::Mutex;
 use super::IdProvider;
 
 /// Provides an implementation of IdProvider that is Send.
-/// Too slow to be of any practical use; for test only
+///
+/// This is probably too slow to be of any practical use; for test only
 #[derive(Debug, Default)]
 pub struct IdProviderMutex {
     internal_counter: Mutex<u64>,
