+++ src/samples/randoms/mod.rs
@@ -1,12 +1,3 @@
-// A thread pool will allow a given resource, identified by an id, to be distributed across many threads
-//
-// The resource must have a defined interface and a defined set of request/response messages.
-//
-// These are things that can run in parallel with their peers and therefore ideally want to be run
-// across several threads.
-//
-// In order to aid testing without taking a dependency on one of these resource a test resource
-// will be implemented here
 pub mod message_processor;
 pub mod randoms_request;
 pub mod randoms_response;
@@ -25,17 +16,17 @@ use {
     randoms_response::{randoms_init_response::RandomsInitResponse, RandomsResponse},
 };
 
-/// This represents a simple "element" which is hosted inside the thread pool
+/// This represents a simple collection of random numbers which is hosted inside the thread pool
 ///
 /// It is tied to a particular thread by the modulus of its id.
 ///
 /// The interface that it supports is governed by its implementation of the ElementProcess trait.
-/// This in turn needs to be supported by the use of two enums of supported Requests and Responses
+/// This in turn needs to be supported by the use of two enums of supported requests and responses
 ///
 /// It supports the following operations
 /// Init       creates a new Random with an stack based store of random numbers
-/// GetMean    calculates the mean of the contained numbers
-/// GetSum     calculates the sum of the contained numbers
+/// Mean    calculates the mean of the contained numbers
+/// Sum     calculates the sum of the contained numbers
 #[derive(Debug, PartialEq, Eq)]
 pub struct Randoms {
     pub id: u64,
