+++ src/samples/randoms/randoms_request/mean_request.rs
@@ -2,6 +2,7 @@ use crate::{id_targeted::IdTargeted, thread_request::ThreadRequest};
 
 use super::RandomsRequest;
 
+/// This is the message sent to request that the Randoms struct (with the given id) calculates the mean of the random numbers it contains.
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct MeanRequest {
     pub id: u64,
