+++ src/samples/randoms/randoms_request/mod.rs
@@ -8,6 +8,12 @@ pub mod mean_request;
 pub mod randoms_init_request;
 pub mod sum_request;
 
+/// This enum defines the full range of request that a Random struct can be sent
+///
+/// This in effects defines the api support by the Random.
+///
+/// Each request is required to produce a response so there is a corresponding
+/// responses enum, which by convention, is similarly named
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub enum RandomsRequest {
     Mean(MeanRequest),
