+++ src/samples/randoms/randoms_response/randoms_init_response.rs
@@ -2,6 +2,7 @@ use crate::{id_targeted::IdTargeted, thread_response::ThreadResponse};
 
 use super::RandomsResponse;
 
+/// The response from a request to create a new Randoms struct within the thread pool
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct RandomsInitResponse {
     pub id: u64,
