+++ src/samples/randoms_batch/randoms_batch_request/randoms_batch_init_request.rs
@@ -5,6 +5,11 @@ use crate::{
 
 use super::RandomsBatchRequest;
 
+/// This is the request that is sent to create a new RandomsBatch
+/// It contains a field to configure the size of the contained child thread pool.
+/// As the this thread pool is shared it will only ever be used by the first request to create a RandomsBatch
+///
+/// RandomsBatches will also need to share a common "source of ids" for the Randoms that it will create
 #[derive(Debug, PartialEq, Clone)]
 pub struct RandomsBatchInitRequest {
     pub id: u64,
