+++ src/samples/randoms_batch/randoms_batch_request/sum_of_sums_request.rs
@@ -2,6 +2,8 @@ use crate::{id_targeted::IdTargeted, thread_request::ThreadRequest};
 
 use super::RandomsBatchRequest;
 
+/// This is the message that is sent to request that a given RandomsBatch calculates the sum of all of the
+/// sums of its contained Randoms
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct SumOfSumsRequest {
     pub id: u64,
